feat(gui): run filter validation on local filter string
It's slightly slower but provides instant feedback.
